Safari-611.1.10/LayoutTests/svg/animations/multiple-animations-fill-freeze-expected.txt

SVG 1.1 dynamic animation tests


This checks the effect on multiple animations on one target

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect1.x.animVal.value is 0
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 0
PASS rect2.x.baseVal.value is 0
PASS rect3.x.animVal.value is 0
PASS rect3.x.baseVal.value is 0
PASS rect1.x.animVal.value is 50
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 50
PASS rect2.x.baseVal.value is 0
PASS rect3.x.animVal.value is 50
PASS rect3.x.baseVal.value is 0
PASS rect1.x.animVal.value is 100
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 100
PASS rect2.x.baseVal.value is 0
PASS rect3.x.animVal.value is 100
PASS rect3.x.baseVal.value is 0
PASS rect1.x.animVal.value is 100
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 0
PASS rect2.x.baseVal.value is 0
PASS rect3.x.animVal.value is 100
PASS rect3.x.baseVal.value is 0
PASS rect1.x.animVal.value is 100
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 0
PASS rect2.x.baseVal.value is 0
PASS rect3.x.animVal.value is 100
PASS rect3.x.baseVal.value is 0
PASS rect1.x.animVal.value is 100
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 0
PASS rect2.x.baseVal.value is 0
PASS rect3.x.animVal.value is 100
PASS rect3.x.baseVal.value is 0
PASS rect1.x.animVal.value is 150
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 150
PASS rect2.x.baseVal.value is 0
PASS rect3.x.animVal.value is 150
PASS rect3.x.baseVal.value is 0
PASS rect1.x.animVal.value is 200
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 200
PASS rect2.x.baseVal.value is 0
PASS rect3.x.animVal.value is 200
PASS rect3.x.baseVal.value is 0
PASS rect1.x.animVal.value is 250
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 250
PASS rect2.x.baseVal.value is 0
PASS rect3.x.animVal.value is 250
PASS rect3.x.baseVal.value is 0
PASS rect1.x.animVal.value is 250
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 250
PASS rect2.x.baseVal.value is 0
PASS rect3.x.animVal.value is 100
PASS rect3.x.baseVal.value is 0
PASS rect1.x.animVal.value is 250
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 250
PASS rect2.x.baseVal.value is 0
PASS rect3.x.animVal.value is 100
PASS rect3.x.baseVal.value is 0
PASS successfullyParsed is true

TEST COMPLETE
